+++ src/main/java/com/java/luckyhankki/dto/product/ProductDetailResponse.java
@@ -1,4 +1,4 @@
-package com.java.luckyhankki.dto;
+package com.java.luckyhankki.dto.product;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import io.swagger.v3.oas.annotations.media.Schema;
