+++ src/test/java/com/java/luckyhankki/controller/StoreControllerTest.java
@@ -1,7 +1,7 @@
 package com.java.luckyhankki.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.java.luckyhankki.dto.ProductResponse;
+import com.java.luckyhankki.dto.product.ProductResponse;
 import com.java.luckyhankki.dto.StoreRequest;
 import com.java.luckyhankki.dto.StoreResponse;
 import com.java.luckyhankki.service.ProductService;
